+++ site/src/other.rs
@@ -116,7 +116,7 @@ pub fn StackIdioms() -> impl IntoView {
         <Editor example="[, @A@B]"/>
         <Editor example="[◌ @A@B]"/>
         <Editor example="[,, @A@B]"/>
-        <Editor example="[,: @A@B]"/>
+        <Editor example="[⟜: @A@B]"/>
         <Editor example="[⊙. @A@B]"/>
         <Editor example="[⊙◌ @A@B]"/>
         <Editor example="[⊙: @A@B@C]"/>
@@ -412,6 +412,14 @@ pub fn Combinators() -> impl IntoView {
             ("⊟⊓¯⇌", 2, "D2", "Dovekie"),
         ),
         (
+            view!(<Prim prim=On/>).into_view(),
+            ("⊟⟜+", 2, "N", "Eastern Nicator"),
+        ),
+        (
+            view!(<Prim prim=Over/>).into_view(),
+            ("⊟+,", 2, "ν", "Western Nicator"),
+        ),
+        (
             view!(<Prim prim=Dip/>).into_view(),
             ("⊟⊙+", 3, "E", "Eagle"),
         ),
@@ -426,7 +434,7 @@ pub fn Combinators() -> impl IntoView {
         ),
         (
             view!(<Prim prim=Bracket/>).into_view(),
-            ("⊟⊓¯+", 3, "P", "Eastern Parotia"),
+            ("⊟⊓¯+", 3, "R", "Eastern Parotia"),
         ),
         (
             view!(<Prim prim=Bracket/>).into_view(),
@@ -459,11 +467,11 @@ pub fn Combinators() -> impl IntoView {
                 }
             }
             let diagram = format!("/combinators/{symbol}.svg");
-            let note = ["X", "χ", "P", "ρ"].contains(&symbol).then(|| {
+            let note = ["N", "ν", "X", "χ", "R", "ρ"].contains(&symbol).then(|| {
                 view! {
                     <sup>" "<span
                         style="text-decoration: underline dotted; font-size: 0.8em; cursor: help;"
-                        title="X, χ, P, and ρ are not standard named combinators. They are included here because Uiua can express them easily.">
+                        title="N, ν, X, χ, R, and ρ are not standard named combinators. They are included here because Uiua can express them easily.">
                         "*"
                     </span></sup>
                 }